Zip: Move tests to separate files
